Return correct error: coordinatorConnection
This was causing a panic when a user group did not exist. The connection
loop checks for a non-nil `resp.Err`, and if found then returns the
local variable `err` which is nil.

Later on this nil value is used to verify that the connection is non-nil
and therefore causes panic when the connection is expanded.