Fix consumer failover
There were two cases (that I found) where bad error handling when
talking to a bad broker.

If a broker is completely down but we think it's the leader for the
partition, yet we don't have a cached connection to it, then we'll keep
looping trying to establish a connection instead of refreshing the
metadata and noticing that there's a different leader we should talk to.

The second case occurs with protocol errors in fetch responses. The
protocol errors (such as NotLeaderForPartition) are attached to each
partition in the response. But the consumer code was only considering
and handling these errors as error return values to c.conn.Fetch(); when
NotLeaderForPartition was encountered in part.Err it was just returned
directly to the user and subsequent calls to Consume() would just keep
yielding the same error. Furthermore, when handling these errors, in
addition to refreshing metadata, we need to discard the cached
connection so that on the next retry we consult the (possibly updated)
metadata to learn which broker to use.

Fixes #27.