Remove now-duplicated response-trimming code from Consumer
When fetching compressed messages, Kafka can return messages that
precede the requested offset (because it must return an entire
compressed batch at once).

Removing leading unrequested messages was already implemented in the
Consumer loop, but I didn't notice this when I implemented compressed
messages. I added this functionality again in the lower-level Fetch
method of the connection.

I decided to remove the original copy of the code and leave the trimming
logic in the connection code. I think this is the best place:

* It's a low-level protocol detail; without this the Consume() code
  seems to operate at a more consistent level of abstraction.
* The trimming cannot happen any lower (in the protocol handling code)
  without making ReadFetchResp be aware of the request (so we know what
  offset was requested), but that doesn't seem right.