when fetching, copy highwater mark (tip offset)
When fetching messages, partition highwater offset (tip offset) is returned
within response as well. Copy it to every message fetched in batch as
TipOffset attribute.

While the value might be a bit missleading because of messages caching (tip
offset is property of the whole fetch request, not single message), to not
break alreayd existing API, value is rewritten to all hosted messages.

For slow consumers this may result in big differences between
message.TipOffset and *current* highwater mark value of partition.