Add AllowTopicCreation
This adds a broker configuration option to enable automatic topic
creation. This defaults to false.

This is useful in clusters where you have a dumb proxy that is
forwarding messages into Kafka which has topic creation enabled. Note
that this is not the fastest of operations, it is effectively a metadata
miss, so requires at least two server roundtrips.