Clean up logging
This is a BREAKING CHANGE. It removes the deprecated Logger and also
consolidates everybody to use a global log objection. This fixes some
unplumbed logging (connection) and also makes it easy to capture all
logging in the test suite plus to have one logging object to control in
a production application.