allow creating topics on new Kafka brokers with empty metadata
When a new Kafka broker is started, it returns no metadata.
So Dial() returned an error when that was the case.

The code has been updated so that Dial() does not return an error
anymore when the broker has empty metadata.
Produce.Produce() has also been updated to create a topic when
metadata is empty.

Fixes #58