Fix Consumer leader change failure
If the leader for a partition becomes a follower for that partition,
then a fetch request returns ErrNotLeaderForPartition. The Consumer
properly handled this case and refreshed metadata.

If the leader for a partition hands off the partition entirely and is no
longer a broker for that partition, it will instead return
ErrUnknownTopicOrPartition. In this case, the Consumer returned the
error and never refreshed metadata, which means that the Consumer is
stuck.

This problem can be reproduced by creating a Kafka cluster with 2 nodes.
Make a single topic with 1 partition and 1 replica. Start consuming,
then force a leader change.