add simple long polling for fetch request
If there is no messages to return fetch handler will lock
until next Produce call.

Probably the best way to solve this problem would be creating
a queue per topic, however considering simplicity of this commit and the
fact that this is a testing server I've decided to go with this solution.