preliminary support for multiple record batches
When sending multiple messages, kafka > 0.11 uses a repetition structure
that was not clear from the protocol documentation.  We are seeing one
message per record batch, with multiple record batches.

This is not a normal array, we should just keep trying to find batches
until we hit the limit of message set size.